Made a component out of user actions so that it could be placed on the landing page, kept on the user settings and used on its' own pgae for usability. New translations will be required.
