+++ src/pages/Actions.js
@@ -1,277 +1,22 @@
-import React, { useEffect, useState } from 'react';
-import SubmittedActionCard from '../components/SubmittedActionCard';
+import React from 'react';
 import {
   Box,
-  Button,
-  Stack,
   Typography,
-  Tab,
 } from '@mui/material';
-import { TabPanel, TabContext, TabList } from '@mui/lab';
-import { API } from 'aws-amplify';
-import {
-  getAllSubmittedActionsForUser,
-  getSingleUser,
-} from '../graphql/queries';
-import useMediaQuery from '@mui/material/useMediaQuery';
+import useTranslation from '../components/customHooks/translations';
+import UserActions from '../components/UserActions';
 
-const Actions = ({ databaseUser, setUser, userType }) => {
-  const [showMore, setShowMore] = useState({
-    validated: false,
-    unvalidated: false,
-    failed: false,
-  });
-  const [validatedActions, setValidatedActions] = useState();
-  const [unvalidatedActions, setUnvalidatedActions] = useState();
-  const [failedActions, setFailedActions] = useState();
-  const tabs = ['Validated', 'Awaiting Validation', 'Did Not Pass Validation'];
-  const [selectedTab, setSelectedTab] = useState(tabs[0]);
-
-  const scrollableTabs = useMediaQuery((theme) => theme.breakpoints.down('sm'));
-
-  useEffect(() => {
-    if (databaseUser) {
-      getUserActions(databaseUser.user_id);
-    }
-  }, [databaseUser]);
-
-  const getCurrentDatabaseUser = async (id) => {
-    const userRes = await API.graphql({
-      query: getSingleUser,
-      variables: { user_id: id },
-    });
-    setUser(userRes.data.getSingleUser);
-  };
-
-  const getUserActions = async (id) => {
-    const actionRes = await API.graphql({
-      query: getAllSubmittedActionsForUser,
-      variables: { user_id: id },
-    });
-    const allActions = actionRes.data.getAllSubmittedActionsForUser;
-    //filter for all validated actions
-    const validated = allActions.filter((action) => action.is_validated);
-    setValidatedActions(validated);
-    //filter for all unvalidated actions
-    const unvalidated = allActions.filter(
-      (action) => !action.is_validated && !action.is_rejected
-    );
-    setUnvalidatedActions(unvalidated);
-    //filter for all failed actions
-    const failed = allActions.filter((action) => action.is_rejected);
-    setFailedActions(failed);
-  };
-
-  const updateValidatedActions = async () => {
-    const actionRes = await API.graphql({
-      query: getAllSubmittedActionsForUser,
-      variables: { user_id: databaseUser.user_id },
-    });
-    const allActions = actionRes.data.getAllSubmittedActionsForUser;
-    //filter for all validated actions
-    const validated = allActions.filter((action) => action.is_validated);
-    setValidatedActions(validated);
-  };
-
-  const renderValidatedActionCards = () => {
-    //return if validatedActions is not null or undefined and contains at least 1 item
-    if (Array.isArray(validatedActions) && validatedActions.length !== 0) {
-      return (
-        <Box
-          sx={{
-            height: '110vh',
-            overflow: 'auto',
-            padding: '0.25em',
-          }}
-        >
-          <Stack spacing={2}>
-            {showMore.validated
-              ? validatedActions.map((action, index) => (
-                  <SubmittedActionCard
-                    key={index}
-                    action={action}
-                    showUnapproveButton={userType === 'Admin'}
-                    getActions={updateValidatedActions}
-                  />
-                ))
-              : validatedActions
-                  .slice(0, 3)
-                  .map((action, index) => (
-                    <SubmittedActionCard
-                      key={index}
-                      action={action}
-                      showUnapproveButton={userType === 'Admin'}
-                      getActions={updateValidatedActions}
-                    />
-                  ))}
-            {validatedActions.length > 3 && (
-              <Button
-                sx={{ mt: '3em' }}
-                variant="outlined"
-                onClick={() =>
-                  setShowMore((prev) => ({
-                    ...prev,
-                    validated: !showMore.validated,
-                  }))
-                }
-              >
-                View {showMore.validated ? 'Less' : 'More'}
-              </Button>
-            )}
-          </Stack>
-        </Box>
-      );
-    } else {
-      return (
-        <Box sx={{ textAlign: 'center' }}>
-          <Typography variant="subtitle2">
-            There are no validated actions to show
-          </Typography>
-        </Box>
-      );
-    }
-  };
-
-  const renderUnvalidatedActionCards = () => {
-    //return if unvalidatedActions is not null or undefined and contains at least 1 item
-    if (Array.isArray(unvalidatedActions) && unvalidatedActions.length !== 0) {
-      return (
-        <Box
-          sx={{
-            height: '110vh',
-            overflow: 'auto',
-            padding: '0.25em',
-          }}
-        >
-          <Stack spacing={2}>
-            {showMore.unvalidated
-              ? unvalidatedActions.map((action, index) => (
-                  <SubmittedActionCard key={index} action={action} />
-                ))
-              : unvalidatedActions
-                  .slice(0, 3)
-                  .map((action, index) => (
-                    <SubmittedActionCard key={index} action={action} />
-                  ))}
-            {unvalidatedActions.length > 3 && (
-              <Button
-                sx={{ mt: '3em' }}
-                variant="outlined"
-                onClick={() =>
-                  setShowMore((prev) => ({
-                    ...prev,
-                    unvalidated: !showMore.unvalidated,
-                  }))
-                }
-              >
-                View {showMore.unvalidated ? 'Less' : 'More'}
-              </Button>
-            )}
-          </Stack>
-        </Box>
-      );
-    } else {
-      return (
-        <Box sx={{ textAlign: 'center' }}>
-          <Typography variant="subtitle2">
-            There are no unvalidated actions to show
-          </Typography>
-        </Box>
-      );
-    }
-  };
-
-  const renderFailedActionCards = () => {
-    //return if failedActions is not null or undefined and contains at least 1 item
-    if (Array.isArray(failedActions) && failedActions.length !== 0) {
-      return (
-        <Box
-          sx={{
-            height: '110vh',
-            overflow: 'auto',
-            padding: '0.25em',
-          }}
-        >
-          <Stack spacing={2}>
-            {showMore.failed
-              ? failedActions.map((action, index) => (
-                  <SubmittedActionCard key={index} action={action} />
-                ))
-              : failedActions
-                  .slice(0, 3)
-                  .map((action, index) => (
-                    <SubmittedActionCard key={index} action={action} />
-                  ))}
-            {failedActions.length > 3 && (
-              <Button
-                sx={{ mt: '3em' }}
-                variant="outlined"
-                onClick={() =>
-                  setShowMore((prev) => ({
-                    ...prev,
-                    failed: !showMore.failed,
-                  }))
-                }
-              >
-                View {showMore.failed ? 'Less' : 'More'}
-              </Button>
-            )}
-          </Stack>
-        </Box>
-      );
-    } else {
-      return (
-        <Box sx={{ textAlign: 'center' }}>
-          <Typography variant="subtitle2">
-            There are no failed actions to show
-          </Typography>
-        </Box>
-      );
-    }
-  };
-
-  const handleTabChange = (e, newValue) => {
-    setSelectedTab(newValue);
-  };
-
+const Actions = ({ databaseUser}) => {
+  const translation = useTranslation();
   return (
     <>
       {databaseUser && (
         <>
           <Box sx={{ textAlign: { xs: 'center', md: 'left' } }}>
             <Typography variant="h1" sx={{ m: '0 0 1.25em' }}>
-              My Logged Actions
+              {translation.myActions}
             </Typography>
-            <TabContext value={selectedTab}>
-              <Box
-                sx={{
-                  borderBottom: 1,
-                  borderTop: 1,
-                  borderColor: 'divider',
-                  maxWidth: { xs: 320, sm: '100%' },
-                }}
-              >
-                <TabList
-                  onChange={handleTabChange}
-                  aria-label="view user action tabs"
-                  scrollButtons
-                  allowScrollButtonsMobile
-                  variant={scrollableTabs ? 'scrollable' : 'fullWidth'}
-                  centered={!scrollableTabs}
-                >
-                  <Tab label={tabs[0]} value={tabs[0]} />
-                  <Tab label={tabs[1]} value={tabs[1]} />
-                  <Tab label={tabs[2]} value={tabs[2]} />
-                </TabList>
-              </Box>
-              <TabPanel value={tabs[0]}>
-                {renderValidatedActionCards()}
-              </TabPanel>
-              <TabPanel value={tabs[1]}>
-                {renderUnvalidatedActionCards()}
-              </TabPanel>
-              <TabPanel value={tabs[2]}>{renderFailedActionCards()}</TabPanel>
-            </TabContext>
+            <UserActions />
           </Box>
         </>
       )}
